Move stacktrace helpers to new script
