+++ tests/test_utils.py
@@ -4,8 +4,7 @@ import json
 from datetime import datetime
 
 from crashsimilarity import utils
-from crashsimilarity.utils import StackTraceProcessor
-from crashsimilarity.utils import StackTracesGetter
+from crashsimilarity.stacktrace_utils import StackTracesGetter, StackTraceProcessor
 
 
 class UtilsTest(unittest.TestCase):
